feat: terms, 레이아웃 재활용 목적의 tabs 수정
